Update changed return types in Amiga library
FossilOrigin-Name: cc34546854c9bc53a5baa84bbf2b7aac68b26ee1b6e9a50d0974866fac237d60

+++ src/amiga-glue.m
@@ -370,7 +370,7 @@ glue_of_socket_address_equal PPC_PARAMS(const of_socket_address_t *address1,
 	return of_socket_address_equal(address1, address2);
 }
 
-uint32_t __saveds
+unsigned long __saveds
 glue_of_socket_address_hash PPC_PARAMS(const of_socket_address_t *address)
 {
 	M68K_ARG(const of_socket_address_t *, address, a0)
